feat :
playerData와 weaponData 골드,코스트 비교하여 계산하는 방향 구현

+++ Assets/Scripts/Weapon/weaponData.cs
@@ -9,7 +9,7 @@ public class WeaponData : ScriptableObject
     [Header("기본 속성")]
     public float Attack;
     public float Critical;
-    public int NeedGold;
+    public float NeedGold;
     public string WeaponName;
     public Sprite WeaponIcon;
